Refactore ElasticWait
