+++ perl/src/main/java/io/perl/impl/PerformanceRecorder.java
@@ -15,10 +15,10 @@ import io.perl.PerlConfig;
 import io.perl.PerlPrinter;
 import io.perl.TimeStamp;
 import io.time.Time;
-
 import javax.annotation.concurrent.NotThreadSafe;
-import java.util.concurrent.locks.LockSupport;
 
+
+@NotThreadSafe
 public final class PerformanceRecorder {
     final private int windowIntervalMS;
     final private int idleNS;
@@ -37,7 +37,7 @@ public final class PerformanceRecorder {
 
     public void run(final long secondsToRun, final long totalRecords) {
         final long msToRun = secondsToRun * Time.MS_PER_SEC;
-        final ElasticWaitCounter idleCounter = new ElasticWaitCounter(windowIntervalMS, idleNS,
+        final ElasticWait idleCounter = new ElasticWait(windowIntervalMS, idleNS,
                 Math.min(windowIntervalMS, PerlConfig.DEFAULT_TIMEOUT_MS));
         final long startTime = time.getCurrentTime();
         boolean doWork = true;
@@ -84,9 +84,9 @@ public final class PerformanceRecorder {
                             periodicRecorder.stopWindow(ctime);
                             periodicRecorder.startWindow(ctime);
                             idleCounter.reset();
-                            idleCounter.setElasticCount(diffTime);
+                            idleCounter.setElastic(diffTime);
                         } else {
-                            idleCounter.updateElasticCount(diffTime);
+                            idleCounter.updateElastic(diffTime);
                         }
                     }
                 }
@@ -99,48 +99,4 @@ public final class PerformanceRecorder {
         PerlPrinter.log.info("Performance Recorder Exited");
     }
 
-
-    /**
-     * Private class for counter implementation to reduce time.getCurrentTime() invocation.
-     */
-    @NotThreadSafe
-    final static private class ElasticWaitCounter {
-        final private int windowInterval;
-        final private int idleNS;
-        final private double countRatio;
-        final private long minIdleCount;
-        private long elasticCount;
-        private long idleCount;
-        private long totalCount;
-
-        public ElasticWaitCounter(int windowInterval, int idleNS, int timeoutMS) {
-            this.windowInterval = windowInterval;
-            this.idleNS = idleNS;
-            countRatio = (Time.NS_PER_MS * 1.0) / this.idleNS;
-            minIdleCount = (long) (countRatio * timeoutMS);
-            elasticCount = minIdleCount;
-            idleCount = 0;
-            totalCount = 0;
-        }
-
-        public boolean waitAndCheck() {
-            LockSupport.parkNanos(idleNS);
-            idleCount++;
-            totalCount++;
-            return idleCount > elasticCount;
-        }
-
-        public void reset() {
-            idleCount = 0;
-        }
-
-        public void updateElasticCount(long diffTime) {
-            elasticCount = Math.max((long) (countRatio * (windowInterval - diffTime)), minIdleCount);
-        }
-
-        public void setElasticCount(long diffTime) {
-            elasticCount = (totalCount * windowInterval) / diffTime;
-            totalCount = 0;
-        }
-    }
 }
